scanner: post-scan cards refactor
