+++ packages/scanner/package.json
@@ -4,7 +4,7 @@
     "private": true,
     "homepage": "/scanner",
     "scripts": {
-        "dev": "PORT=3001 BROWSER=none CLIENT_ENV=development craco start --verbose",
+        "dev": "cross-env PORT=3001 BROWSER=none CLIENT_ENV=development craco start --verbose",
         "build": "CLIENT_ENV=production craco build --verbose",
         "lint:css": "stylelint ./src/**/*.css",
         "lint-fix:css": "stylelint --fix ./src/**/*.css",
