+++ packages/scanner/src/app/app.tsx
@@ -9,8 +9,7 @@ import '~/shared/lib/date';
 import { AppProvider } from '~/model/app-provider/app-provider';
 import { useCheckVersion } from '~/shared/hooks/use-check-version';
 import { Screens } from '~/app/screens';
-
-import { ViewProvider } from '../model/view-provider/view-provider';
+import { ViewProvider } from '~/model/view-provider';
 
 const ErrorFallback: FC<FallbackProps> = ({ error, resetErrorBoundary }) => (
     <div role='alert'>
@@ -22,7 +21,7 @@ const ErrorFallback: FC<FallbackProps> = ({ error, resetErrorBoundary }) => (
 
 const App: FC = () => {
     useCheckVersion();
-
+    console.log('app');
     return (
         // @ts-ignore
         <ErrorBoundary fallback={ErrorFallback as ReactElement}>
