+++ packages/scanner/src/components/misc/misc.module.css
@@ -1,10 +1,3 @@
-.card {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    width: 100vw;
-}
-
 .cardDefault {
     display: flex;
     align-content: space-around;
@@ -15,13 +8,6 @@
 
     button {
         height: 100%;
-        padding: 1em;
-    }
-}
-
-.anoncard {
-    button {
-        width: 50vw;
         padding: 1em;
     }
 }
