+++ packages/scanner/src/components/post-scan-cards/ui/card-container/card-container.tsx
@@ -1,0 +1,7 @@
+import css from './card-container.module.css';
+
+export const CardContainer = (props) => {
+    const { children } = props;
+
+    return <div className={css.cardContainer}>{children}</div>;
+};
