+++ packages/scanner/src/components/post-scan/post-scan-anon/post-scan-anon.tsx
@@ -1,20 +1,16 @@
 import type { FC } from 'react';
-import { useEffect } from 'react';
 
-import { AppColor } from '~/model/app-provider/app-provider';
-import { GreenAnonCard } from '~/components/misc';
 import { useApp } from '~/model/app-provider';
+import { FeedAnonCard } from '~/components/post-scan-cards/feed-anon-card/feed-anon-card';
 
 import { useFeedVol } from '../post-scan.utils';
 
 export const PostScanAnon: FC<{
     closeFeed: () => void;
 }> = ({ closeFeed }) => {
-    const { kitchenId, mealTime, setColor } = useApp();
+    const { kitchenId, mealTime } = useApp();
 
     const [doFeed] = useFeedVol(undefined, mealTime, closeFeed, kitchenId);
 
-    useEffect(() => setColor(AppColor.GREEN), []);
-
-    return <GreenAnonCard close={closeFeed} doFeed={doFeed} />;
+    return <FeedAnonCard close={closeFeed} doFeed={doFeed} />;
 };
