+++ packages/scanner/src/components/post-scan/post-scan-vol/post-scan-vol.tsx
@@ -2,8 +2,10 @@ import type { FC } from 'react';
 import { useLiveQuery } from 'dexie-react-hooks';
 
 import { FeedType, type Volunteer } from '~/db';
-import { ErrorMsg, GreenCard, YellowCard } from '~/components/misc';
 import { AppColor, useApp } from '~/model/app-provider';
+import { WarningCard } from '~/components/post-scan-cards/warning-card';
+import { FeedCard } from '~/components/post-scan-cards/feed-card';
+import { ErrorCard } from '~/components/post-scan-cards/error-card';
 
 import { getTodayStart, getVolTransactionsAsync, useFeedVol, validateVol } from '../post-scan.utils';
 
@@ -19,7 +21,7 @@ export const PostScanVol: FC<{
     const [doFeed, doNotFeed] = useFeedVol(vol, mealTime, closeFeed, kitchenId);
 
     if (!volTransactions) {
-        return <ErrorMsg close={closeFeed} doNotFeed={doNotFeed} msg={`Бейдж не найден: ${qrcode}`} />;
+        return <ErrorCard close={closeFeed} doNotFeed={doNotFeed} msg={`Бейдж не найден: ${qrcode}`} />;
     }
 
     const { isRed, msg } = validateVol(vol, volTransactions, kitchenId, mealTime!, false);
@@ -27,13 +29,13 @@ export const PostScanVol: FC<{
     if (msg.length > 0) {
         if (isRed) {
             setColor(AppColor.RED);
-            return <ErrorMsg close={closeFeed} doNotFeed={doNotFeed} msg={msg} />;
+            return <ErrorCard close={closeFeed} doNotFeed={doNotFeed} msg={msg} />;
         } else {
             setColor(AppColor.YELLOW);
-            return <YellowCard close={closeFeed} doFeed={doFeed} doNotFeed={doNotFeed} vol={vol} msg={msg} />;
+            return <WarningCard close={closeFeed} doFeed={doFeed} doNotFeed={doNotFeed} vol={vol} msg={msg} />;
         }
     }
 
     setColor(vol.feed_type === FeedType.Child ? AppColor.BLUE : AppColor.GREEN);
-    return <GreenCard close={closeFeed} doFeed={doFeed} vol={vol} />;
+    return <FeedCard close={closeFeed} doFeed={doFeed} vol={vol} />;
 };
