+++ packages/scanner/src/screens/main-screen/main.tsx
@@ -1,10 +1,11 @@
-import React, { useCallback, useContext, useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 
-import { ErrorMsg, LastUpdated } from '~/components/misc/misc';
+import { LastUpdated } from '~/components/misc/misc';
 import { PostScan } from '~/components/post-scan';
 import { db } from '~/db';
 import { Scan } from '~/components/scan/scan';
 import { useApp } from '~/model/app-provider';
+import { ErrorCard } from '~/components/post-scan-cards/error-card';
 
 import css from './main.module.css';
 
@@ -25,7 +26,7 @@ export const MainScreen = React.memo(function MainScreen() {
         <div className={css.main}>
             {!scanResult && <Scan onScan={setScanResult} />}
             {scanResult && <PostScan closeFeed={closeFeedDialog} qrcode={scanResult} />}
-            {appError && <ErrorMsg close={closeFeedDialog} msg={appError} />}
+            {appError && <ErrorCard close={closeFeedDialog} msg={appError} />}
             {/*<LastUpdated count={volCount} ts={lastSyncStart || 0} />*/}
         </div>
     );
